Fixing test

+++ datastr/hashtable_test.go
@@ -11,7 +11,7 @@ func TestNewHashTable(t *testing.T) {
 	ht := NewHashTable(size)
 
 	assert.Equal(t, size, ht.size, "Expected size %d, but got %d", size, ht.size)
-	assert.Equal(t, size-1, ht.sizemask, "Expected sizemask %d, but got %d", size-1, ht.sizemask)
+	assert.Equal(t, uint64(size-1), ht.sizemask, "Expected sizemask %d, but got %d", size-1, ht.sizemask)
 	assert.NotNil(t, ht.table, "Expected table to be initialized, but it's nil")
 	assert.False(t, ht.empty(), "Expected HashTable to be not empty, but it's empty")
 }
@@ -21,7 +21,7 @@ func TestEmptyHashTable(t *testing.T) {
 	ht := NewHashTable(0)
 
 	assert.Equal(t, size, ht.size, "Expected size %d, but got %d", size, ht.size)
-	assert.Equal(t, size, ht.sizemask, "Expected sizemask %d, but got %d", size, ht.sizemask)
+	assert.Equal(t, uint64(size), ht.sizemask, "Expected sizemask %d, but got %d", size, ht.sizemask)
 	assert.Empty(t, ht.table, "Expected table to be nil but it's initialized")
 	assert.True(t, ht.empty(), "Expected HashTable to be empty, but it's not")
 }
